Renamed 'src' to 'app'
